Complete tasks from basic level
Create a basic window with a text in the middle of it, so that at
resizing text is being redrawn.

Usually the target folder is added to `.gitignore` file, however
in case someone has no rust compiler installed I have committed
the executable too so that the program can be checked for working
state.